# project_docs/generated/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CreateInventoryCorrectionData": {
        "properties": {
          "correction_event_id": {
            "title": "Correction Event Id",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "new_quantity": {
            "title": "New Quantity",
            "type": "string"
          }
        },
        "required": [
          "correction_event_id",
          "item_id",
          "new_quantity"
        ],
        "title": "CreateInventoryCorrectionData",
        "type": "object"
      },
      "CreateInventoryCorrectionRequest": {
        "properties": {
          "corrected_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Corrected Quantity"
          },
          "expected_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Expected Quantity"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "expected_quantity",
          "corrected_quantity",
          "reason"
        ],
        "title": "CreateInventoryCorrectionRequest",
        "type": "object"
      },
      "CreateInventoryStockOutData": {
        "properties": {
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "new_quantity": {
            "title": "New Quantity",
            "type": "string"
          },
          "stock_out_event_id": {
            "title": "Stock Out Event Id",
            "type": "string"
          }
        },
        "required": [
          "stock_out_event_id",
          "item_id",
          "new_quantity"
        ],
        "title": "CreateInventoryStockOutData",
        "type": "object"
      },
      "CreateInventoryStockOutRequest": {
        "properties": {
          "expected_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Expected Quantity"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "stock_out_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Stock Out Quantity"
          }
        },
        "required": [
          "item_id",
          "expected_quantity",
          "stock_out_quantity",
          "reason"
        ],
        "title": "CreateInventoryStockOutRequest",
        "type": "object"
      },
...
      "CreateSessionMessageData": {
        "properties": {
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_run_id": {
            "title": "Task Run Id",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "task_run_id",
          "status"
        ],
        "title": "CreateSessionMessageData",
        "type": "object"
      },
      "CreateSessionMessageRequest": {
        "properties": {
          "client_request_id": {
            "title": "Client Request Id",
            "type": "string"
          },
          "media_ids": {
            "items": {
              "type": "string"
            },
            "title": "Media Ids",
            "type": "array"
          },
          "message_type": {
            "title": "Message Type",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "message_type",
          "client_request_id"
        ],
        "title": "CreateSessionMessageRequest",
        "type": "object"
      },
...
      "DataEnvelope_CreateInventoryCorrectionData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateInventoryCorrectionData"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[CreateInventoryCorrectionData]",
        "type": "object"
      },
      "DataEnvelope_CreateInventoryStockOutData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateInventoryStockOutData"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[CreateInventoryStockOutData]",
        "type": "object"
      },
...
      "DataEnvelope_CreateSessionMessageData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateSessionMessageData"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[CreateSessionMessageData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionMessageItem": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "media_ids": {
            "items": {
              "type": "string"
            },
            "title": "Media Ids",
            "type": "array"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "message_type": {
            "title": "Message Type",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "task_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Run Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "message_id",
          "session_id",
          "actor_type",
          "actor_id",
          "message_type",
          "text",
          "media_ids",
          "task_run_id",
          "created_at"
        ],
        "title": "SessionMessageItem",
        "type": "object"
      },
      "SessionMessagesMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "SessionMessagesMeta",
        "type": "object"
      },
      "SessionMessagesResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SessionMessageItem"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/SessionMessagesMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SessionMessagesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Store Manager Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inventory-events/corrections": {
      "post": {
        "operationId": "post_inventory_correction_api_v1_inventory_events_corrections_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CreateInventoryCorrectionData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Inventory Correction",
        "tags": [
          "inventory-events"
        ]
      }
    },
    "/api/v1/inventory-events/stock-out": {
      "post": {
        "operationId": "post_inventory_stock_out_api_v1_inventory_events_stock_out_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryStockOutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CreateInventoryStockOutData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Inventory Stock Out",
        "tags": [
          "inventory-events"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/messages": {
      "get": {
        "operationId": "get_session_messages_api_v1_sessions__session_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMessagesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Messages",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "post_session_message_api_v1_sessions__session_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CreateSessionMessageData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Session Message",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}